+++ app/templates/components/nav-menu.hbs
@@ -1,6 +1,6 @@
 <div class="flex items-center">
-  {{#link-to "docs.index" class="hide-on-tablet"}}Docs{{/link-to}}
-  {{#link-to "docs.page" "tutorials/example-apps" class="ml-3 hide-on-tablet"}}Tutorials{{/link-to}}
+  {{!-- {{#link-to "docs.index" class="hide-on-tablet"}}Docs{{/link-to}} --}}
+  {{!-- {{#link-to "docs.page" "tutorials/example-apps" class="ml-3 hide-on-tablet"}}Tutorials{{/link-to}} --}}
   {{#if session.isAuthenticated}}
     {{user-menu user=user class="ml-3"}}
   {{else}}
