+++ app/templates/components/organizations/integrations/github-enterprise-settings.hbs
@@ -47,6 +47,6 @@
     </form>
   </div>
   <div class="rounded bg-yellow-000 border border-yellow-100 p-2 mb-2">
-    <strong>Reminder:</strong> Each project must be linked to a GitHub repository for <strong>{{#link-to "docs.page" "learn/github-integration"}}Pull Request integration{{/link-to}}</strong> to work. You can verify the link in each project's settings.
+    {{!-- <strong>Reminder:</strong> Each project must be linked to a GitHub repository for <strong>{{#link-to "docs.page" "learn/github-integration"}}Pull Request integration{{/link-to}}</strong> to work. You can verify the link in each project's settings. --}}
   </div>
 </section>
