+++ app/templates/components/project-container.hbs
@@ -68,7 +68,7 @@
         </ol>
 
         <p>
-          {{#link-to "docs.index" class="Button"}}View setup documentation{{/link-to}}
+          {{!-- {{#link-to "docs.index" class="Button"}}View setup documentation{{/link-to}} --}}
         </p>
 
         <h5>Environment variables</h5>
