+++ app/templates/components/supported-cis-image.hbs
@@ -1,3 +1,3 @@
-{{#link-to "docs.index"}}
+{{!-- {{#link-to "docs.index"}} --}}
   <img src="/images/cis.png">
-{{/link-to}}
+{{!-- {{/link-to}} --}}
