+++ app/templates/index.hbs
@@ -93,7 +93,7 @@
     <h3 class="text-xl md:text-2xl mb-1">Simple and fast</h3>
     <p class="text-lg text-gray-400">
       We handle all the heavy lifting of consistent screenshot rendering and visual diffing in our systems, so you don't have to. Percy comes with drop-in open source clients for many Ruby, Python, and JavaScript frameworks, plus a command-line client for static websites.
-      See the {{#link-to "docs.index"}}docs{{/link-to}} to get started.
+      {{!-- See the {{#link-to "docs.index"}}docs{{/link-to}} to get started. --}}
     </p>
   </div>
 
@@ -172,7 +172,7 @@
       <p class="text-gray-400">
         Test your responsive designs—simply provide a list of breakpoint widths and we take care of the rest. See how CSS changes will affect your entire site, including on mobile devices. Catch mobile visual bugs before launch.
         <br>
-        {{#link-to "docs.page" "learn/responsive"}}See docs{{/link-to}}
+        {{!-- {{#link-to "docs.page" "learn/responsive"}}See docs{{/link-to}} --}}
       </p>
       <p>
         <img class="u-capped-width" style="border: 1px solid #ddd" src="/images/example-responsive-375.jpg">
@@ -186,7 +186,7 @@
     <h2>Get started</h2>
     <div class="row">
       <div class="six columns">
-      <h3>{{#link-to "docs.page" "clients/ruby/capybara"}}Rails »{{/link-to}}</h3>
+      {{!-- <h3>{{#link-to "docs.page" "clients/ruby/capybara"}}Rails »{{/link-to}}</h3> --}}
 {{#code-block lang="ruby"}}
   gem 'percy-capybara'
 {{/code-block}}
@@ -203,7 +203,7 @@
 
 {{/code-block}}
 
-      <h3>{{#link-to "docs.page" "clients/ruby/cli"}}Static websites »{{/link-to}}</h3>
+      {{!-- <h3>{{#link-to "docs.page" "clients/ruby/cli"}}Static websites »{{/link-to}}</h3> --}}
 
 {{#code-block lang="bash"}}
   $ gem install percy-cli
@@ -211,7 +211,7 @@
 {{/code-block}}
     </div>
     <div class="six columns">
-      <h3>{{#link-to "docs.page" "clients/javascript/ember"}}Ember »{{/link-to}}</h3>
+      {{!-- <h3>{{#link-to "docs.page" "clients/javascript/ember"}}Ember »{{/link-to}}</h3> --}}
 {{#code-block lang="bash"}}
   $ ember install ember-percy
 {{/code-block}}
