+++ app/templates/organization/project/settings.hbs
@@ -24,7 +24,7 @@
       <div class="bg-white rounded shadow-1 p-2 mb-2">
         <h2 class="text-xl font-semibold mb-1">Browsers</h2>
         <p class="text-gray-400 mb-2">
-          This setting lets you change which browser your snapshots are created in. Better yet, you can enable multiple browsers for cross browser testing snapshots! {{#link-to "docs.page" "learn/cross-browser-visual-testing"}}See docs{{/link-to}}
+          {{!-- This setting lets you change which browser your snapshots are created in. Better yet, you can enable multiple browsers for cross browser testing snapshots! {{#link-to "docs.page" "learn/cross-browser-visual-testing"}}See docs{{/link-to}} --}}
         </p>
         <div class="p-2 bg-gray-000 rounded border">
           {{projects/browser-family-selector
